Updated queries, db schema, added listview for day availability
Query now works when searching for date. Although current date is hardcoded, next commit will take passed in parameters to the query. 

The db schema changed, "isBooked" removed from tutor_sessions table, and added it to availability_time table.

Added listview to the student_single_day_availability.xml layout

When clicking the listview, it brings you to a new fragment with the times for that specific day, which in the next commit will be included.

TODO:
Update query to accept parameters
Swipe-to-delete functionality on the listview for day availability. This requires a query which searches for the specific day, start time, and end time selected.